requirements.txt
google-generativeai
python-dotenv
pillow
markdown
fpdf2
